+++ dsol-core/src/main/java/nl/tudelft/simulation/jstats/distributions/DistNegBinomial.java
@@ -67,7 +67,7 @@ public class DistNegBinomial extends DistDiscrete
 
     /** {@inheritDoc} */
     @Override
-    public double probability(final int observation)
+    public double probability(final long observation)
     {
         if (observation >= 0)
         {
